LayoutTests/editing/pasteboard/paste-blockquote-into-blockquote.html

<html>
<head>
<script src="../../resources/dump-as-markup.js"></script>
<style>
blockquote {
    color: blue;
    border-left: 2px solid blue;
    margin: 0px;
    padding: 0 0 0 20px;
}
</style>
</head>
<body>
<p id="description">This tests pasting a blockquote into a blockquote. Nothing below should be double-blockquoted.</p>
<div id="test" contenteditable="true"><blockquote type='cite' id="block">One</blockquote></div>
<script>
var sel = window.getSelection();
var block = document.getElementById("block");

sel.setPosition(block, 1);
document.execCommand("InsertHTML", false, "<span><blockquote type='cite'><div>Two</div><div>Three</div></blockquote></span>");

Markup.description(document.getElementById('description').textContent);
Markup.dump('test');

</script>
</body>
</html>
