LayoutTests/editing/pasteboard/paste-blockquote-into-blockquote-expected.txt

This tests pasting a blockquote into a blockquote. Nothing below should be double-blockquoted.
| <blockquote>
|   id="block"
|   type="cite"
|   "One"
|   "Two"
| <blockquote>
|   type="cite"
|   <div>
|     "Three<#selection-caret>"
